+++ CHANGES.md
@@ -5,6 +5,8 @@ Changes in 11.8
 ---------------
 - SQL performance: reduce staff table row size (was 3.8KB) & remove HOMEROOM column in rosariosis.sql & rosariosis_mysql.sql
 - Fix rare DB error "null value in column "last_name" violates not-null constraint" in Student.php & User.php
+- Fix SQL error null value in column "student_id" violates not-null constraint in Requests.php
+
 Changes in 11.7.3
 -----------------
 - Fix regression since 11.7.1 typo generating SQL error in CreateParents.php