+++ CHANGES.md
@@ -11,6 +11,10 @@ Changes in 11.8
 - CSS remove forced large font size for Transcripts PDF in Transcripts.php
 - Add & get PROFILE_ID column in GetTeacher.fnc.php & User.fnc.php
 
+Changes in 11.7.4
+-----------------
+- Fix regression since 11.7 no results if more than 1000 fees but 0 payments for timeframe in DailyTransactions.php
+
 Changes in 11.7.3
 -----------------
 - Fix regression since 11.7.1 typo generating SQL error in CreateParents.php